requirements.txt
discord.py
imgurpython
dotenv
nekos.py